requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
blinker==1.9.0
boto3==1.40.33
botocore==1.40.33
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cohere==5.18.0
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
executing==2.2.1
fastavro==1.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.27
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.38.0
googleapis-common-protos==1.70.0
griffe==1.14.0
groq==0.31.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
logfire==4.7.0
logfire-api==4.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.14.0
mdurl==0.1.2
mistralai==1.9.10
multidict==6.6.4
narwhals==2.5.0
nexus-rpc==1.1.0
numpy==2.3.3
openai==1.108.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-ai==1.0.8
pydantic-ai-slim==1.0.8
pydantic-evals==1.0.8
pydantic-graph==1.0.8
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
pyperclip==1.10.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
streamlit==1.49.1
temporalio==1.17.0
tenacity==9.1.2
tokenizers==0.22.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
types-protobuf==6.32.1.20250918
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
